usm_pipe was overwritten but it is shared among all GPUs so that's why the multiGPU runs raised error - The scaling is not good now, as we increase the number of GPUs performance decreases - It is kind of expected because we are decreasing the workload so occupancy decreases and overheads increase maybe we need to play with range4 - reverted uint64_t for the numEdges to uint32_t since we are not using uint64_t datatype anymore - used memset for the initialisation of the ScanFullDevice so that it will get rid of the else if condition in the ExploreNeighbours3

+++ src/host.cpp
@@ -87,7 +87,7 @@ int main(int argc, char * argv[])
   CSRGraph graph_cpu = loadMatrix(1,datasetName);
 
   uint32_t numCols = graph_cpu.meta[1];  // cols -> total number of vertices
-  uint64_t numEdges  = graph_cpu.meta[2];  // nonZ count -> total edges
+  uint32_t numEdges  = graph_cpu.meta[2];  // nonZ count -> total edges
   uint32_t numRows   = graph_cpu.meta[0];  // this it the value we want! (rows)
       std::cout << std::setw(20) << std::left << "- # vertices" << std::setw(20) << numRows << std::endl;
       std::cout << std::setw(20) << std::left << "- # edges" << std::setw(20) << numEdges << std::endl;
